+++ tests/unit_tests/fixtures/substitutions/02-expressions.input.yaml
@@ -9,6 +9,7 @@ substitutions:
   numberOne: 1
   var1: 79
   double_width: ${width * 2}
+  template_expression: !literal ${ width * 3 }
 
 test_list:
   - "The area is ${width * height}"
@@ -25,3 +26,5 @@ test_list:
   - chr(97) = ${ chr(97) }
   - len([1,2,3]) = ${ len([1,2,3]) }
   - width = ${width}, double_width = ${double_width}
+  - ${ template_expression } is literal
+  - "${ eval(template_expression, {'width': 7 }) } == 21"
